backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rc-mail-assistant"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "tenacity>=8.3",
    "structlog>=24.1",
    "python-multipart>=0.0.9",
    "jinja2>=3.1",
    "httpx>=0.27",
    "pydantic-settings>=2.2",
    "importlib-metadata>=7.0",
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "google-auth-httplib2>=0.1",
    "html2text>=2024.2",
    "apscheduler>=3.10,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
